Reverted to Set. Added randomize(). Updated JavaDocs
I realized this morning the correctElement into the Set was bad coupling. The implementer should be left to decide how the incorrectElements and correctElement should used. This also removed the issue of maintaining order in the Set. Reverted to HashSet.

Added a randomization method called randomize(). This can be called to update a new answer without requiring instantiation again. Better design.